browser/components/firefoxview/tests/chrome/test_fxview_search_textbox.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>FxviewSearchTextbox Tests</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <link rel="localization" href="browser/firefoxView.ftl"/>
  <script type="module" src="chrome://browser/content/firefoxview/fxview-search-textbox.mjs"></script>
</head>
<body>
<p id="display"></p>
<div id="content">
  <fxview-search-textbox></fxview-search-textbox>
</div>
<pre id="test"></pre>
<script>
  const { BrowserTestUtils } = ChromeUtils.importESModule(
    "resource://testing-common/BrowserTestUtils.sys.mjs"
  );

  add_task(async function test_search_and_clear() {
    const component = document.querySelector("fxview-search-textbox");

    info("Enter a search query.");
    let promiseEvent = BrowserTestUtils.waitForEvent(
      component,
      "fxview-search-textbox-query"
    );
    synthesizeMouseAtCenter(component, {});
    sendChar("Q");
    let query = (await promiseEvent).detail.query;
    is(query, "Q", "Event was fired with the correct query.");

    info("Clear the search query.");
    promiseEvent = BrowserTestUtils.waitForEvent(
      component,
      "fxview-search-textbox-query"
    );
    synthesizeMouseAtCenter(component.clearButton, {});
    query = (await promiseEvent).detail.query;
    is(query, "", "Event was fired with a blank query.");
  });
</script>
</body>
</html>
